@@ -367,7 +367,7 @@ namespace DataStructures.Tests
         }
 
         [Test]
-        public static void TestCompareThrowsException()
+        public static void ArraysOfDifferentLengthsAreNotEqual()
         {
             // Arrange
             var testObj1 = new BitArray("110");
@@ -376,23 +376,7 @@ namespace DataStructures.Tests
             // Act
 
             // Assert
-            _ = Assert.Throws<Exception>(() => Assert.IsTrue(testObj1 == testObj2));
-        }
-
-        [Test]
-        public static void TestCompareTo()
-        {
-            // Arrange
-            var testObj1 = new BitArray("110");
-            var testObj2 = new BitArray("110");
-            var testObj3 = new BitArray("100");
-
-            // Act
-
-            // Assert
-            Assert.AreEqual(testObj1.CompareTo(testObj3), 1);
-            Assert.AreEqual(testObj3.CompareTo(testObj1), -1);
-            Assert.AreEqual(testObj1.CompareTo(testObj2), 0);
+            Assert.False(testObj1 == testObj2);
         }
 
         #endregion COMPARE TESTS
